# docs/openapi.yaml
openapi: 3.0.4
info:
  title: Meethub MVP Backend API
  description: >
    OpenAPI specification for the TikTok-like MVP backend. The gateway entry
    point is Traefik on port 8080. Direct service ports are also listed for
    local development and debugging.
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Gateway
  - url: http://localhost:8001
    description: Auth service
  - url: http://localhost:8002
    description: Upload service
  - url: http://localhost:8003
    description: Feed service
  - url: http://localhost:8004
    description: Moderation service
...
paths:
...
  /upload/request:
    post:
      tags: [upload]
      summary: Create upload session and generate presigned URL
      operationId: requestUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
      responses:
        '200':
          description: Upload session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /upload/image/request:
    post:
      tags: [upload]
      summary: Create image upload session and generate presigned URL
      operationId: requestImageUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageUploadRequest'
      responses:
        '200':
          description: Image upload session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /upload/complete:
    post:
      tags: [upload]
      summary: Complete upload and enqueue transcoding
      operationId: completeUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '202':
          description: Upload accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/image/complete:
    post:
      tags: [upload]
      summary: Complete image upload and publish image item
      operationId: completeImageUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteImageUploadRequest'
      responses:
        '200':
          description: Image uploaded and published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /upload/status/{uploadId}:
    get:
      tags: [upload]
      summary: Get upload processing status
      operationId: getUploadStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UploadIdPath'
      responses:
        '200':
          description: Upload status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /internal/uploads/{uploadId}/status:
    put:
      tags: [internal]
      summary: Update upload status from worker
      operationId: internalUpdateUploadStatus
      security:
        - internalApiKey: []
      parameters:
        - $ref: '#/components/parameters/UploadIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalUpdateUploadStatusRequest'
      responses:
        '200':
          description: Upload status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
  parameters:
    UploadIdPath:
      in: path
      name: uploadId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object
      required:
        - detail
...
    StatusMessage:
      type: object
      additionalProperties:
        type: string
...
    UploadStatus:
      type: string
      enum:
        - pending
        - uploaded
        - transcoding
        - ready
        - error
    UploadRequest:
      type: object
      properties:
        file_name:
          type: string
        content_type:
          type: string
          default: video/mp4
      required:
        - file_name
        - content_type
    ImageUploadRequest:
      type: object
      properties:
        file_name:
          type: string
        content_type:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          default: image/jpeg
      required:
        - file_name
        - content_type
    UploadSessionResponse:
      type: object
      properties:
        upload_id:
          type: string
          format: uuid
        upload_url:
          type: string
        s3_key:
          type: string
        expires_in_seconds:
          type: integer
          default: 1800
      required:
        - upload_id
        - upload_url
        - s3_key
        - expires_in_seconds
    ImageUploadSessionResponse:
      type: object
      properties:
        upload_id:
          type: string
          format: uuid
        upload_url:
          type: string
        image_url:
          type: string
        s3_key:
          type: string
        expires_in_seconds:
          type: integer
          default: 1800
      required:
        - upload_id
        - upload_url
        - image_url
        - s3_key
        - expires_in_seconds
    CompleteUploadRequest:
      type: object
      properties:
        upload_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 5000
          default: ""
        hashtags:
          type: array
          items:
            type: string
        location:
          $ref: '#/components/schemas/LocationPayload'
      required:
        - upload_id
    CompleteImageUploadRequest:
      type: object
      properties:
        upload_id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 5000
          default: ""
        hashtags:
          type: array
          items:
            type: string
        location:
          $ref: '#/components/schemas/LocationPayload'
      required:
        - upload_id
    UploadStatusResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/UploadStatus'
        description:
          type: string
          nullable: true
        location:
          $ref: '#/components/schemas/LocationPayload'
        error_message:
          type: string
          nullable: true
        hashtags:
          type: array
          items:
            type: string
        image_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - status
        - hashtags
        - created_at
    InternalUpdateUploadStatusRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/UploadStatus'
        error_message:
          type: string
          nullable: true
      required:
        - status
...
    LocationPayload:
      type: object
      properties:
        name:
          type: string
          nullable: true
          maxLength: 255
        city:
          type: string
          nullable: true
          maxLength: 128
        latitude:
          type: number
          format: double
          nullable: true
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          nullable: true
          minimum: -180
          maximum: 180
...